adding repository and service classes. closes #13

+++ fraud-service/src/main/java/com/luv4code/fraud/app/web/ui/model/FraudCheckHistory.java
@@ -18,8 +18,8 @@ public class FraudCheckHistory {
     @Id
     @SequenceGenerator(name = "fraud_id_sequence", sequenceName = "fraud_id_sequence")
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "fraud_id_sequence")
-    private Long id;
-    private Long customerId;
+    private Integer id;
+    private Integer customerId;
     private Boolean isFraudster;
     private LocalDateTime createdAt;
 }
